accuracy attempts
the pitch estimation is way out, usually anywhere from 4-10Hz. Sometimes it kicks up the freq to like 200 above at the end of the note. I have added a  simple noise filter in attempts to increase the accuracy, and I have also changed the algorithm. Need to mess around with values to find the most accurate setup now