update
Changed fonts to new monospaced ones.
Updated colours, fonts and the theme.
New reference pitch page added.
Fiddled with the layout of the main page a bit.
Made some functions more reusable.
Might have to move the new ref page to a new file and when I get navigation working.

Next is to the ViewModel or something else working to get the data displayed properly. This will require new stuff, but also changing around the current code to get the stuff to update and display the current info.