Holy Jesus, it works. Thank the lord.
Well. Finally got the UI updating with the values that are calculated. Only took like a week of research to figure out a solution. Only went through like 7 different solutions that all didn't work. Phind is kind of a godsend bruh like damn. But anyway, shit works now.

It uses an object, which can be updated with the calculated variables. This is then observed by the ViewModel. This ViewModel is then observed by the UI, which recomposes upon a change to the ViewModel. Most jank and roundabout way of doing it but whatever, its the only thing I could think of that could possibly work after exhausting everything else. That and some sort of global variables.