Name and Icon change
Changed the name and the icon of the app.

Not happy with it, but the idea I came up with, I just couldn't make bring it to fruition, so used a name idea that my mum got from Chat GPT.

Not happy, but its something. At least its mediocre, not non-existent or shit.