Shit finally works
Minimum product reached.
Only need to add some comments to the code to fully describe it.
Also want to add a small change to the bar which will add a sharp and flat symbol at the appropriate ends of the bar.

If I've got time and feel like it, I could try and learn some animations and animate the bar and progress indicator between state to make it look nicer

Thank god its pretty much over. I enjoyed making it but it has always felt like a struggle.